% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{atlas1006}
\alias{atlas1006}
\title{HITChip Atlas with 1006 Western Adults}
\format{
The dataset in \code{TreeSummarizedExperiment} format.
}
\usage{
data(atlas1006)
}
\value{
Loads the dataset in R.
}
\description{
This dataset contains genus-level microbiota profiling with
HITChip for 1006 western adults with no reported health complications,
reported in Lahti et al. (2014) \url{https://doi.org/10.1038/ncomms5344}.
}
\details{
The dataset is also available via the microbiome R package
\url{http://microbiome.github.com/microbiome} in \code{phyloseq} format.
}
\references{
Lahti et al. Nature Communications 5:4344, 2014.
\url{https://doi.org/10.1038/ncomms5344}
}
\author{
Huang Lin \email{huanglinfrederick@gmail.com}
}
\keyword{data}
