% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dietswap}
\alias{dietswap}
\title{Diet Swap Data}
\format{
The dataset in \code{TreeSummarizedExperiment} format.
}
\usage{
data(dietswap)
}
\value{
Loads the dataset in R.
}
\description{
The diet swap dataset represents a study with African
and African American groups undergoing a two-week diet swap.
For details, see
\url{https://doi.org/10.1038/ncomms7342}.
}
\details{
The dataset is also available via the microbiome R package
\url{http://microbiome.github.com/microbiome} in \code{phyloseq} format.
}
\references{
O'Keefe et al. Nature Communications 6:6342, 2015.
\url{https://doi.org/10.1038/ncomms7342}
}
\author{
Huang Lin \email{huanglinfrederick@gmail.com}
}
\keyword{data}
