% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secom_dist.R
\name{secom_dist}
\alias{secom_dist}
\title{Sparse estimation of distance correlations among microbiomes}
\usage{
secom_dist(
  data,
  taxa_are_rows = TRUE,
  assay.type = assay_name,
  assay_name = "counts",
  rank = tax_level,
  tax_level = NULL,
  aggregate_data = NULL,
  meta_data = NULL,
  pseudo = 0,
  prv_cut = 0.5,
  lib_cut = 1000,
  corr_cut = 0.5,
  wins_quant = c(0.05, 0.95),
  R = 1000,
  thresh_hard = 0,
  max_p = 0.005,
  n_cl = 1,
  verbose = TRUE
)
}
\arguments{
\item{data}{a \code{list} of the input data.
The \code{data} parameter should be either a
\code{matrix}, \code{data.frame}, \code{phyloseq} or a \code{TreeSummarizedExperiment}
object. Both \code{phyloseq} and \code{TreeSummarizedExperiment} objects
consist of a feature table (microbial count table), a sample metadata table,
a taxonomy table (optional), and a phylogenetic tree (optional).
If a \code{matrix} or \code{data.frame} is provided, ensure that the row
names of the \code{metadata} match the sample names (column names if
\code{taxa_are_rows} is TRUE, and row names otherwise) in \code{data}.
if a \code{phyloseq} or a \code{TreeSummarizedExperiment} is used, this
standard has already been enforced. For detailed information, refer to
\code{?phyloseq::phyloseq} or
\code{?TreeSummarizedExperiment::TreeSummarizedExperiment}.
It is recommended to use low taxonomic levels, such as OTU or species level,
as the estimation of sampling fractions requires a large number of taxa.
If working with multiple ecosystems, such as gut and tongue, stack the data
by specifying the list of input data as
\code{data = list(gut = pseq1, tongue = pseq2)}.}

\item{taxa_are_rows}{logical. Whether taxa are positioned in the rows of the
feature table. Default is TRUE.}

\item{assay.type}{alias for \code{assay_name}.}

\item{assay_name}{character. Name of the count table in the data object
(only applicable if data object is a \code{(Tree)SummarizedExperiment}).
Default is "counts".
See \code{?SummarizedExperiment::assay} for more details.}

\item{rank}{alias for \code{tax_level}.}

\item{tax_level}{character. The taxonomic level of interest. The input data
can be agglomerated at different taxonomic levels based on your research
interest. Default is NULL, i.e., do not perform agglomeration, and the
SECOM analysis will be performed at the lowest taxonomic level of the
input \code{data}.}

\item{aggregate_data}{The abundance data that has been aggregated to the desired
taxonomic level. This parameter is required only when the input data is in
\code{matrix} or \code{data.frame} format. For \code{phyloseq} or \code{TreeSummarizedExperiment}
data, aggregation is performed by specifying the \code{tax_level} parameter.}

\item{meta_data}{a \code{data.frame} containing sample metadata.
This parameter is mandatory when the input \code{data} is a generic
\code{matrix} or \code{data.frame}. Ensure that the row names of the \code{metadata} match the
sample names (column names if \code{taxa_are_rows} is TRUE, and row names
otherwise) in \code{data}.}

\item{pseudo}{numeric. Add pseudo-counts to the data.
Default is 0 (no pseudo-counts).}

\item{prv_cut}{a numerical fraction between 0 and 1. Taxa with prevalences
(the proportion of samples in which the taxon is present)
less than \code{prv_cut} will be excluded in the analysis. For example,
if there are 100 samples, and a taxon has nonzero counts present in less than
100*prv_cut samples, it will not be considered in the analysis.
Default is 0.50.}

\item{lib_cut}{a numerical threshold for filtering samples based on library
sizes. Samples with library sizes less than \code{lib_cut} will be
excluded in the analysis. Default is 1000.}

\item{corr_cut}{numeric. To avoid false positives caused by taxa with small
variances, taxa with Pearson correlation coefficients greater than
\code{corr_cut} with the estimated sample-specific bias will be flagged.
When taxa are flagged, the pairwise correlation coefficient between them will
be set to 0s. Default is 0.5.}

\item{wins_quant}{a numeric vector of probabilities with values between
0 and 1. Replace extreme values in the abundance data with less
extreme values. Default is \code{c(0.05, 0.95)}. For details,
see \code{?DescTools::Winsorize}.}

\item{R}{numeric. The number of replicates in calculating the p-value for
distance correlation. For details, see \code{?energy::dcor.test}.
Default is 1000.}

\item{thresh_hard}{Numeric. Pairwise correlation coefficients
(in their absolute value) that are less than or equal to \code{thresh_hard}
will be set to 0. Default is 0, i.e. do not apply hard thresholding.}

\item{max_p}{numeric. Obtain the sparse correlation matrix by
p-value filtering. Pairwise correlation coefficients with p-value greater
than \code{max_p} will be set to 0s. Default is 0.005.}

\item{n_cl}{numeric. The number of nodes to be forked. For details, see
\code{?parallel::makeCluster}. Default is 1 (no parallel computing).}

\item{verbose}{logical. Whether to display detailed progress messages.
Default is TRUE.}
}
\value{
a \code{list} with components:
        \itemize{
        \item{ \code{s_diff_hat}, a numeric vector of estimated
        sample-specific biases.}
        \item{ \code{y_hat}, a matrix of bias-corrected abundances}
        \item{ \code{mat_cooccur}, a matrix of taxon-taxon co-occurrence
        pattern. The number in each cell represents the number of complete
        (nonzero) samples for the corresponding pair of taxa.}
        \item{ \code{dcorr}, the sample distance correlation matrix
        computed using the bias-corrected abundances \code{y_hat}.}
        \item{ \code{dcorr_p}, the p-value matrix corresponding to the sample
        distance correlation matrix \code{dcorr}.}
        \item{ \code{dcorr_fl}, the sparse correlation matrix obtained by
        p-value filtering \code{dcorr} based on the cutoff specified in
        \code{max_p}, followed by hard thresholding at \code{thresh_hard}.}
        }
}
\description{
Obtain the sparse correlation matrix for distance correlations
between taxa.
}
\details{
The \href{https://doi.org/10.1214/009053607000000505}{distance correlation},
which is a measure of dependence between two random variables, can be used to
quantify any dependence, whether linear, monotonic, non-monotonic or
nonlinear relationships.
}
\examples{
library(ANCOMBC)
if (requireNamespace("microbiome", quietly = TRUE)) {
    data(atlas1006, package = "microbiome")
    # subset to baseline
    pseq = phyloseq::subset_samples(atlas1006, time == 0)

    # run secom_dist function
    set.seed(123)
    res_dist = secom_dist(data = list(pseq), taxa_are_rows = TRUE,
                          tax_level = "Phylum",
                          aggregate_data = NULL, meta_data = NULL, pseudo = 0,
                          prv_cut = 0.5, lib_cut = 1000, corr_cut = 0.5,
                          wins_quant = c(0.05, 0.95), R = 1000,
                          thresh_hard = 0.3, max_p = 0.005, n_cl = 2)

    dcorr_fl = res_dist$dcorr_fl
} else {
    message("The 'microbiome' package is not installed. Please install it to use this example.")
}

}
\seealso{
\code{\link{secom_linear}} \code{\link{data_sanity_check}}
}
\author{
Huang Lin
}
