% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ancom.R
\name{ancom}
\alias{ancom}
\title{Analysis of Composition of Microbiomes (ANCOM)}
\usage{
ancom(
  phyloseq,
  p_adj_method = "holm",
  prv_cut = 0.1,
  lib_cut = 0,
  main_var,
  adj_formula = NULL,
  rand_formula = NULL,
  lme_control = NULL,
  struc_zero = FALSE,
  neg_lb = FALSE,
  alpha = 0.05,
  n_cl = 1
)
}
\arguments{
\item{phyloseq}{a phyloseq-class object, which consists of a feature table
(microbial observed abundance table), a sample metadata, a taxonomy table
(optional), and a phylogenetic tree (optional). The row names of the
metadata must match the sample names of the feature table, and the row names
of the taxonomy table must match the taxon (feature) names of the feature
table. See \code{?phyloseq::phyloseq} for more details.}

\item{p_adj_method}{character. method to adjust p-values. Default is "holm".
Options include "holm", "hochberg", "hommel", "bonferroni", "BH", "BY",
"fdr", "none". See \code{?stats::p.adjust} for more details.}

\item{prv_cut}{a numerical fraction between 0 and 1. Taxa with prevalences
less than \code{prv_cut} will be excluded in the analysis. Default
is 0.10.}

\item{lib_cut}{a numerical threshold for filtering samples based on library
sizes. Samples with library sizes less than \code{lib_cut} will be
excluded in the analysis. Default is 0, i.e. do not discard any sample.}

\item{main_var}{character. The name of the main variable of interest.}

\item{adj_formula}{character string representing the formula for
covariate adjustment. Default is \code{NULL}.}

\item{rand_formula}{character string representing the formula for random
effects. For details, see \code{?nlme::lme}. Default is \code{NULL}.}

\item{lme_control}{a list specifying control values for \code{lme} fit.
For details, see ?\code{nlme::lmeControl}. Default is \code{NULL}.}

\item{struc_zero}{logical. whether to detect structural zeros based on
\code{main_var}. \code{main_var} should be discrete. Default is FALSE.}

\item{neg_lb}{logical. whether to classify a taxon as a structural zero using
its asymptotic lower bound. Default is FALSE.}

\item{alpha}{numeric. level of significance. Default is 0.05.}

\item{n_cl}{numeric. The number of nodes to be forked. For details, see
\code{?parallel::makeCluster}. Default is 1 (no parallel computing).}
}
\value{
a \code{list} with components:
        \itemize{
        \item{ \code{res},  a \code{data.frame} containing ANCOM
        result for the variable specified in \code{main_var},
        each column is:}
        \itemize{
        \item{ \code{W}, test statistics.}
        \item{ \code{detected_0.9, detected_0.8, detected_0.7, detected_0.6},
        logical vectors representing whether a taxon is differentially
        abundant under a series of cutoffs. For example, TRUE in
        \code{detected_0.7} means the number of ALR transformed models where
        the taxon is differentially abundant with regard to the main variable
        outnumbers \code{0.7 * (n_taxa - 1)}. \code{detected_0.7} is commonly
        used. Choose \code{detected_0.8} or \code{detected_0.9} for more
        conservative results, or choose \code{detected_0.6} for more liberal
        results.}
        }
        \item{ \code{zero_ind}, a logical \code{matrix} with TRUE indicating
        the taxon is identified as a structural zero for the specified
        main variable.}
        \item{ \code{beta_data}, a numeric \code{matrix} containing pairwise
        coefficients for the main variable of interest in ALR transformed
        regression models.}
        \item{ \code{p_data}, a numeric \code{matrix} containing pairwise
        p-values for the main variable of interest in ALR transformed
        regression models.}
        \item{ \code{q_data}, a numeric \code{matrix} containing adjusted
        p-values by applying the \code{p_adj_method} to the \code{p_data}
        matrix.}
        }
}
\description{
Determine taxa whose absolute abundances, per unit volume, of
the ecosystem (e.g. gut) are significantly different with changes in the
covariate of interest (e.g. group). The current version of
\code{ancom} function implements ANCOM in cross-sectional and longitudinal
datasets while allowing for covariate adjustment.
}
\details{
The definition of structural zero can be found at
\href{https://doi.org/10.3389/fmicb.2017.02114}{ANCOM-II}.
Setting \code{neg_lb = TRUE} indicates that you are using both criteria
stated in section 3.2 of
\href{https://doi.org/10.3389/fmicb.2017.02114}{ANCOM-II}
to detect structural zeros; otherwise, the algorithm will only use the
equation 1 in section 3.2 for declaring structural zeros. Generally, it is
recommended to set \code{neg_lb = TRUE} when the sample size per group is
relatively large (e.g. > 30).
}
\examples{
library(microbiome)
library(tidyverse)
data(dietswap)

# Subset to baseline
pseq = subset_samples(dietswap, timepoint == 1)
# Aggregate to family level
family_data = aggregate_taxa(pseq, "Family")

# Run ancombc function
set.seed(123)
out = ancom(phyloseq = family_data,  p_adj_method = "holm",
            prv_cut = 0.10, lib_cut = 0, main_var = "nationality",
            adj_formula = "bmi_group",
            rand_formula = NULL, lme_control = NULL,
            struc_zero = TRUE, neg_lb = TRUE, alpha = 0.05, n_cl = 2)

res = out$res

}
\references{
\insertRef{mandal2015analysis}{ANCOMBC}
}
\seealso{
\code{\link{ancombc}}
}
\author{
Huang Lin
}
