% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ancombc.R
\name{ancombc}
\alias{ancombc}
\title{Analysis of Compositions of Microbiomes with Bias Correction
(ANCOM-BC)}
\usage{
ancombc(
  phyloseq,
  formula,
  p_adj_method = "holm",
  prv_cut = 0.1,
  lib_cut = 0,
  group = NULL,
  struc_zero = FALSE,
  neg_lb = FALSE,
  tol = 1e-05,
  max_iter = 100,
  conserve = FALSE,
  alpha = 0.05,
  global = FALSE
)
}
\arguments{
\item{phyloseq}{a phyloseq-class object, which consists of a feature table
(microbial observed abundance table), a sample metadata, a taxonomy table
(optional), and a phylogenetic tree (optional). The row names of the
metadata must match the sample names of the feature table, and the row names
of the taxonomy table must match the taxon (feature) names of the feature
table. See \code{?phyloseq::phyloseq} for more details.}

\item{formula}{the character string expresses how the microbial absolute
abundances for each taxon depend on the variables in metadata.}

\item{p_adj_method}{character. method to adjust p-values. Default is "holm".
Options include "holm", "hochberg", "hommel", "bonferroni", "BH", "BY",
"fdr", "none". See \code{?stats::p.adjust} for more details.}

\item{prv_cut}{a numerical fraction between 0 and 1. Taxa with prevalences
less than \code{prv_cut} will be excluded in the analysis. Default
is 0.10.}

\item{lib_cut}{a numerical threshold for filtering samples based on library
sizes. Samples with library sizes less than \code{lib_cut} will be
excluded in the analysis. Default is 0, i.e. do not discard any sample.}

\item{group}{character. the name of the group variable in metadata.
\code{group} should be discrete. Specifying \code{group} is required for
detecting structural zeros and performing global test.}

\item{struc_zero}{logical. whether to detect structural zeros based on
\code{group}. Default is FALSE.}

\item{neg_lb}{logical. whether to classify a taxon as a structural zero using
its asymptotic lower bound. Default is FALSE.}

\item{tol}{numeric. the iteration convergence tolerance for the E-M
algorithm. Default is 1e-05.}

\item{max_iter}{numeric. the maximum number of iterations for the E-M
algorithm. Default is 100.}

\item{conserve}{logical. whether to use a conservative variance estimator for
the test statistic. It is recommended if the sample size is small and/or
the number of differentially abundant taxa is believed to be large.
Default is FALSE.}

\item{alpha}{numeric. level of significance. Default is 0.05.}

\item{global}{logical. whether to perform global test. Default is FALSE.}
}
\value{
a \code{list} with components:
        \itemize{
        \item{ \code{feature_table}, a \code{data.frame} of pre-processed
        (based on \code{prv_cut} and \code{lib_cut}) microbial observed
        abundance table.}
        \item{ \code{zero_ind}, a logical \code{matrix} with TRUE indicating
        the taxon is identified as a structural zero for the specified
        \code{group} variable.}
        \item{ \code{samp_frac}, a numeric vector of estimated sampling
        fractions in log scale (natural log). Note that for each sample,
        if it contains missing values for any variable specified in the
        \code{formula}, the corresponding sampling fraction estimate
        for this sample will be \code{NA} since the sampling fraction
        is not estimable with the presence of missing values.}
        \item{ \code{resid}, a \code{matrix} of residuals from the ANCOM-BC
        log-linear (natural log) model.
        Rows are taxa and columns are samples.}
        \item{ \code{delta_em}, estimated sample-specific biases
        through E-M algorithm.}
        \item{ \code{delta_wls}, estimated sample-specific biases through
        weighted least squares (WLS) algorithm.}
        \item{ \code{res},  a \code{list} containing ANCOM-BC primary result,
        which consists of:}
        \itemize{
        \item{ \code{lfc}, a \code{data.frame} of log fold changes
        obtained from the ANCOM-BC log-linear (natural log) model.}
        \item{ \code{se}, a \code{data.frame} of standard errors (SEs) of
        \code{lfc}.}
        \item{ \code{W}, a \code{data.frame} of test statistics.
        \code{W = lfc/se}.}
        \item{ \code{p_val}, a \code{data.frame} of p-values. P-values are
        obtained from two-sided Z-test using the test statistic \code{W}.}
        \item{ \code{q_val}, a \code{data.frame} of adjusted p-values.
        Adjusted p-values are obtained by applying \code{p_adj_method}
        to \code{p_val}.}
        \item{ \code{diff_abn}, a logical \code{data.frame}. TRUE if the
        taxon has \code{q_val} less than \code{alpha}.}
        }
        \item{ \code{res_global},  a \code{data.frame} containing ANCOM-BC
        global test result for the variable specified in \code{group},
        each column is:}
        \itemize{
        \item{ \code{W}, test statistics.}
        \item{ \code{p_val}, p-values, which are obtained from two-sided
        Chi-square test using \code{W}.}
        \item{ \code{q_val}, adjusted p-values. Adjusted p-values are
        obtained by applying \code{p_adj_method} to \code{p_val}.}
        \item{ \code{diff_abn}, A logical vector. TRUE if the taxon has
        \code{q_val} less than \code{alpha}.}
        }
        }
}
\description{
Determine taxa whose absolute abundances, per unit volume, of
the ecosystem (e.g. gut) are significantly different with changes in the
covariate of interest (e.g. group). The current version of
\code{ancombc} function implements Analysis of Compositions of Microbiomes
with Bias Correction (ANCOM-BC) in cross-sectional data while allowing
for covariate adjustment.
}
\details{
The definition of structural zero can be found at
\href{https://doi.org/10.3389/fmicb.2017.02114}{ANCOM-II}.
Setting \code{neg_lb = TRUE} indicates that you are using both criteria
stated in section 3.2 of
\href{https://doi.org/10.3389/fmicb.2017.02114}{ANCOM-II}
to detect structural zeros; otherwise, the algorithm will only use the
equation 1 in section 3.2 for declaring structural zeros. Generally, it is
recommended to set \code{neg_lb = TRUE} when the sample size per group is
relatively large (e.g. > 30).
}
\examples{
#================Build a Phyloseq-Class Object from Scratch==================
library(phyloseq)

otu_mat = matrix(sample(1:100, 100, replace = TRUE), nrow = 10, ncol = 10)
rownames(otu_mat) = paste0("taxon", 1:nrow(otu_mat))
colnames(otu_mat) = paste0("sample", 1:ncol(otu_mat))


meta = data.frame(group = sample(LETTERS[1:4], size = 10, replace = TRUE),
                  row.names = paste0("sample", 1:ncol(otu_mat)),
                  stringsAsFactors = FALSE)

tax_mat = matrix(sample(letters, 70, replace = TRUE),
                 nrow = nrow(otu_mat), ncol = 7)
rownames(tax_mat) = rownames(otu_mat)
colnames(tax_mat) = c("Kingdom", "Phylum", "Class", "Order",
                      "Family", "Genus", "Species")

OTU = otu_table(otu_mat, taxa_are_rows = TRUE)
META = sample_data(meta)
TAX = tax_table(tax_mat)
physeq = phyloseq(OTU, META, TAX)

#========================Run ANCOMBC Using a Real Data=======================

library(microbiome)
library(tidyverse)
data(dietswap)

# Subset to baseline
pseq = subset_samples(dietswap, timepoint == 1)
# Aggregate to family level
family_data = aggregate_taxa(pseq, "Family")

# Run ancombc function
out = ancombc(phyloseq = family_data, formula = "bmi_group + nationality",
              p_adj_method = "holm", prv_cut = 0.10, lib_cut = 1000,
              group = "bmi_group", struc_zero = TRUE, neg_lb = FALSE,
              tol = 1e-5, max_iter = 100, conserve = TRUE,
              alpha = 0.05, global = TRUE)

res = out$res
res_global = out$res_global

}
\references{
\insertRef{kaul2017analysis}{ANCOMBC}

\insertRef{lin2020analysis}{ANCOMBC}
}
\seealso{
\code{\link{ancom}}
}
\author{
Huang Lin
}
