% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secom_dist.R
\name{secom_dist}
\alias{secom_dist}
\title{Sparse estimation of distance correlations among microbiomes}
\usage{
secom_dist(
  pseqs,
  pseudo = 0,
  prv_cut = 0.5,
  corr_cut = 0.5,
  lib_cut = 1000,
  wins_quant = c(0.05, 0.95),
  R = 1000,
  thresh_hard = 0,
  max_p = 0.005,
  n_cl = 1
)
}
\arguments{
\item{pseqs}{a list of phyloseq-class objects. For one single ecosystem,
specify it as \code{pseqs = list(c(phyloseq1, phyloseq2))}, where
\code{phyloseq1} (typically in low taxonomic levels, such as OTU or species
level) is used to estimate biases, while \code{phyloseq2} (can be in any
taxonomic level) is used to compute the correlation matrix.
For multiple ecosystems, simply stack the phyloseq objects. For example,
for two ecosystems (such as gut and tongue), specify it as
\code{pseqs = list(gut = c(phyloseq1, phyloseq2),
tongue = c(phyloseq3, phyloseq4))}.}

\item{pseudo}{numeric. Add pseudo-counts to the data.
Default is 0 (no pseudo-counts).}

\item{prv_cut}{a numerical fraction between 0 and 1. Taxa with prevalences
less than \code{prv_cut} will be excluded in the analysis. Default is 0.5.}

\item{corr_cut}{numeric. To prevent false positives due to taxa with
small variances, taxa with Pearson correlation coefficients greater than
\code{corr_cut} with the estimated sample-specific bias will be flagged.
The pairwise correlation coefficient between flagged taxa will be set to 0s.
Default is 0.5.}

\item{lib_cut}{a numerical threshold for filtering samples based on library
sizes. Samples with library sizes less than \code{lib_cut} will be
excluded in the analysis. Default is 1000.}

\item{wins_quant}{a numeric vector of probabilities with values between
0 and 1. Replace extreme values in the abundance data with less
extreme values. Default is \code{c(0.05, 0.95)}. For details,
see \code{?DescTools::Winsorize}.}

\item{R}{numeric. The number of replicates in calculating the p-value for
distance correlation. For details, see \code{?energy::dcor.test}.
Default is 1000.}

\item{thresh_hard}{Numeric. Set a hard threshold for the correlation matrix.
Pairwise distance correlation less than or equal to \code{thresh_hard}
will be set to 0. Default is 0 (No ad-hoc hard thresholding).}

\item{max_p}{numeric. Obtain the sparse correlation matrix by
p-value filtering. Pairwise correlation coefficient with p-value greater than
\code{max_p} will be set to 0. Default is 0.005.}

\item{n_cl}{numeric. The number of nodes to be forked. For details, see
\code{?parallel::makeCluster}. Default is 1 (no parallel computing).}
}
\value{
a \code{list} with components:
        \itemize{
        \item{ \code{s_diff_hat}, a numeric vector of estimated
        sample-specific biases.}
        \item{ \code{y_hat}, a matrix of bias-corrected abundances}
        \item{ \code{mat_cooccur}, a matrix of taxon-taxon co-occurrence
        pattern. The number in each cell represents the number of complete
        (nonzero) samples for the corresponding pair of taxa.}
        \item{ \code{dcorr}, the sample distance correlation matrix
        computed using the bias-corrected abundances \code{y_hat}.}
        \item{ \code{dcorr_p}, the p-value matrix corresponding to the sample
        distance correlation matrix \code{dcorr}.}
        \item{ \code{dcorr_fl}, the sparse correlation matrix obtained by
        p-value filtering based on the cutoff specified in \code{max_p}.}
        }
}
\description{
Obtain the sparse correlation matrix for distance correlations
between taxa.
}
\details{
The \href{https://projecteuclid.org/journals/annals-of-statistics/volume-35/issue-6/Measuring-and-testing-dependence-by-correlation-of-distances/10.1214/009053607000000505.full}{distance correlation},
which is a measure of dependence between two random variables, can be used to
quantify any dependence, whether linear, monotonic, non-monotonic or
nonlinear relationships.
}
\examples{
library(microbiome)
library(tidyverse)
data(dietswap)

# Subset to baseline
pseq = subset_samples(dietswap, timepoint == 1)
# Genus level data
phyloseq1 = pseq
# Phylum level data
phyloseq2 = aggregate_taxa(pseq, level = "Phylum")

# print(phyloseq1)
# print(phyloseq2)

set.seed(123)
res_dist = secom_dist(pseqs = list(c(phyloseq1, phyloseq2)), pseudo = 0,
                      prv_cut = 0.5, lib_cut = 1000, corr_cut = 0.5,
                      wins_quant = c(0.05, 0.95), R = 1000,
                      thresh_hard = 0.3, max_p = 0.005, n_cl = 1)

dcorr_fl = res_dist$dcorr_fl

}
\seealso{
\code{\link{secom_linear}}
}
\author{
Huang Lin
}
